Add middleware interface tests
We can use pytest-only features, so we don't need the unittest.main stuff.
I tried to remove unittest entirely (no parent class), but assertLogs is much
easier to use than the pytest equivalent, it appears. We might want the
activator tests to be pure pytest because there are useful Flask fixtures.