Ingest raws into group-specific runs.
Putting the raws in uniquely named collections prevents dataset collisions in
our test environment (where the same exposure may be ingested as part of
different groups) and makes it easier to clean up the central repository if we
want to start anew. <instrument>/raw/all still exists as a chained collection,
so any code that assumes its existence should have the expected behavior.