Lower severity of "no images downloaded" failure mode to error.
The "fatal" Python logging level is translated to Google's "critical",
which is defined as "cause more severe problems or outages". While it's
certainly bad if a visit can't be processed at all, it won't affect the
rest of the system.

Google's "error" is defined merely as "likely to cause problems", and
is more appropriate.