Fix bug with multiple exports of same exposure ID.
This bug is currently caused by repetition of input exposure IDs, but
it may also happen in the future as a result of retries, especially
late-stage ones. Solution is to have the export code check for and only
export the latest run, though this comes at the cost of not being able
to retry previously failed exports.