Use generated exposure in uploaded files.
This change makes it possible to upload the same visits repeatedly without
confusing the Butler. However, repeated uploads will still break the APDB,
which assumes that all visits' *timestamps* are unique. Faking the timestamps
in a self-consistent fashion is difficult and dangerous, especially if any
pipeline code uses the times to derive other quantities.