Use findFirst=True in querying templates
Using findFirst so that we can store multiple versions of templates
in one chain collection and let butler search for the dataset to use.

New collections with the updated templates can be added to the chain
collection.  The order in the chain collection determines the query
results, like the usual butler convention.

findFirst is not used in querying calibration because findFirst query
in CALIBRATION-type collection is not supported yet.