Ingest all raws into <instrument>/raw/all.
Previously, there was a single raw collection for each group, and
raw/all was a chained collection that linked them together. This
architecture not only deviated from repository conventions, it had the
same scaling problems as timestamped output runs. Now that raws are
guaranteed to have unique IDs in simulated data, we can safely put
them all in a single standard run.