Add refreshes for central butler to MWI.
While we shouldn't need refresh for the local repo, because there is
only one Butler for it, the central repo has one Butler per worker, and
any individual worker may be long- or short-lived. While it's not a
proper synchronization API, call `central_butler.registry.refresh`
before operations that read from or write to the central repo.