Export the calibration collection
Previously the top-level calibration collection (e.g. "HSC/calib")
in the test repo has been a CALIBRATION collection, so it got
exported with other CALIBRATION collections in the above lines.

In the new repo, the top-level calibration collection will be a
CHAINED collections of selected CALIBRATION collections, so it
needs to be exported separatedly.