Export the entire chain collection structure
Extra collections in the chain in the central repo used to cause
MissingCollectionError in preparing the local butler if some collections
are not necessary for the incoming visit. Here we just export the
entire default chain even if some collections are empty and not useful.
Typically the default chain include refcats, calibration, skymap, etc.