Chain the templates collection to the default collection
While we can ensure that the templates' collection exists in the
central repo, it may not be chained the default collection.
Currently that is the case in most shared repos. So we export it
separately and chain it in the local repo.

This is a workaround until the prompt processing has its own
default top-level collection in the shared repos.