Send visit-level next-visit message
As the fan-out service is standing up, the next visit messages
sent by the upload tester scripts should be at the visit level
as those from the summit, and they are to be consumed by the
fan-out service. A SummitVisit class is added for the summit-style
messages, which have some extra private_ fields.