Fix bad default in MiddlewareInterface._get_output_run.
The intended behavior was that _get_output_run would use the (UTC) date
it was called on, but the use of a default argument caused it to always
use the date the pod was booted on. This led to inconsistent behavior,
since the same visit might be processed by pods of different ages.