Revamp raw.py interface.
This allows LSST cameras to use different path patterns from other
cameras, and it allows LSST cameras to get their group information from
a "sidecar" JSON file rather than the path.

Determining based on a group id whether the images associated with it
have already arrived works for non-LSST cameras but is left for a future
ticket for the LSST cameras.