Remove dedicated output chain from MiddlewareInterface.
The output chain, which combined an output run, raw inputs, and a
defaults that excluded raws, was originally based on the ap_verify
dataset organization, which in turn imitated the collection style
created by pipetask. However, a distinction between raw and default
collections will not be needed in production, and the pipetask layout
is difficult to reconcile with a workflow that may use a different run
for every visit.

The output runs were removed from the "output" collection on e38bda3.
This commit removes the output collection itself, replacing it with
direct use of the defaults collection. make_local_repo and prep_butler
now guarantee that the run collection is part of the default chain in
the dev environment.

The removal of the <instrument>/prompt chain is a behavioral change,
though one that should only affect the local repo and not the
central repo.