Allow non-tenant user for dev testing at USDF
This is to work with the Ceph object store at USDF. The buckets for
Prompt Processing development (s3://rubin-pp and s3://rubin-pp-users)
are tenant buckets while the buckets storing the summit data
(s3://rubin-summit and s3://rubin-summit-users) are not.
A new non-tenant user is created to access all these buckets (DM-38923).
But the tenant bucket names need the tenant prefix, and boto bucket
name validation needs to be disabled.