Pre-emptively turn off rbClassify in AP pipeline.
rbClassify is not yet part of the root ApPipe.yaml, but we don't want
to run it until we have appropriate support (either weights datasets
in our central repos, or a copy of a weights package inside the
container). This support will be added after we get R/B working for
regular ApPipe runs.