Refresh the unit test data repo with ap_verify_ci_hits2015
The unit test central repo was made using ap_verify_ci_hits2015
data repo. DM-38449 changed its skymap and templates.
DM-38853 removed the renaming of the catalogs to gaia and panstarrs
and insertion of the refcats into a specific run.
Here we rebuild the unit test data repo with the current
ap_verify_ci_hits2015 and update the parameters accordingly.