Allow JSON encoding of common types.
It's easy for an add_context decorator to add types that are not JSON
serializable, and the decorator's usage should not depend on the choice
of formatter. This commit makes the formatter more flexible about its
inputs, so that e.g. a set is serialized as an array (the fact that
Grafana would read it as a list is acceptable).