Change group_base and last_group to be a str
A number had been used as the group ID in the HSC tester
for convenience, but the group ID in the nextVisit events
is a string. Currently a timestamp-like string is used in
the nextVisit events from LSST summit.

This commit makes some "group" to be a string, as a preparation
step to make the upload.py script work with LSST data.