Change the file paths of the unobserved HSC files
These 10 files s3://rubin-pp-users/unobserved/HSC/
are used in upload.py HSC tester.

As the group IDs in HSC testers have changed, change
the file paths to be consistent.  Practically this
is not necessary because the unobserved visits'
group IDs do not matter to the generated group IDs.
But this is done just so to be consistent.