Reorganize raw-listener loop.
This change makes "has not timed out" part of the loop condition,
ensuring that the loop will abort in a timely manner even if irrelevant
messages continue being received. The loop body has also been rewritten
to better account for the presence of irrelevant messages.

It's still possible for the activator to wait for up to 2*timeout, but
this will only happen if the message stream stops completely.