Do not raise if export_outputs doesn't match.
Originally, export_outputs would only be called if the caller knew
there were datasets to export, so not finding any files implied the
data ID was wrong. Now that export_outputs may be called when the
pipeline failed at the beginning, this gives false positives. The
original message remains as a warning so that humans can check if it
came up unexpectedly.