Use find_dataset instead of queryDatasetAssociations to filter calibs.
queryDatasetAssociations is extremely slow (the loop over dataset types
takes a minute to complete), so its use makes filtering a bottleneck.
Use find_dataset instead to select by timestamp.

We cannot yet use find_dataset to search for calibs directly, because
it can only be run on one dataset type at a time and we don't yet have
the sophistication to organically determine which types the pipelines
need.