Update the seed config for dev butler central repo
We try to configure the dev repo as similarly as prod
for which the central repo is the shared /repo/embargo.
It is configured for the live ObsTAP service.

As bahaviors can be different based on the repo configs,
it has happened that bugs only appeared with /repo/embargo
as the central repo, but not with a repo without live ObsTAP
config. That led to bugs not shown in dev but in prod.