Catch MissingDatasetTypeError in making quantum graph
If a dataset type isn't known to the local repo, the quantum
graph building now fails with MissingDatasetTypeError. It
was ignored in the past. This scenario is possible in cases
such as a fresh local repo is used and the incoming image
doesn't have matching templates. We want it to fall back to
try the next pipeline. If the repo knows the template's
dataset type but no matching templates exist, it still
generate an empty quantum graph.