Preload machine learning model(s) in prep_butler.
A model is quite large (hundreds of MB), but because the same model is
used regardless of data ID, it should be a one-time overhead for
the pod.

This commit also chains the dummy model to the test repo's
DECam/defaults, something that should have been done when it was added.