Add instructions for distinguishing major and minor releases.
In semantic versioning, compatibility is always defined with respect to
a specific API. In this case, we require compatibility in incoming and
outgoing message schemas, as well as the dimensions version of the
central Butler repo. We do not consider compatibility with respect to
specific tasks or pipelines, partly because Prompt Processing should be
agnostic to those and partly because such changes are nearly impossible
to regulate or track.