Use local calibration collections instead of central ones.
Faithfully reproducing the central calib collections is computationally
expensive. Instead, we create our own calib collections which are tagged
not by the CP products issue, but by the (processing) day on which the
pod first needed to download them. Since the original download checks
for both priority and validity, the local collections impose the same
precedence as the original ones.

All this assumes, however, that we never try to process observations
from previous nights.