Add safeguard to dimension export to central repo.
To avoid potential deadlocks from trying to export all dimensions in
parallel, we instead export a hardcoded list of dimensions. This commit
adds a guard that ignores any dimensions that don't exist in the
current Butler version.

This guard is exercised by our unit tests, which use a version 4 repo.