Parameterize instrument in prepare_one_visit
Although not always true, assuming raw data are part of the
{instrument}/raw/all collection is usually fine.

For HSC-RC2, the raw data in the "HSC/RC2/defaults" collection
(the HSC/raw/RC2 collection) is a subset of the raw data in
the "HSC/raw/all" collection.

The assumption is good enough for now, so I'm defering further
refactoring until the need arises.

Also, query with visit instead of exposure for generality.
Currently either works, but eventually we may get snaps.