Clarify how minor versions work for APDB.
Minor version updates to the APDB or Apdb implementations are
"backwards compatible", but this is used to mean that new clients can
use old databases, not the other way around. This commit rewords the
versioning guidelines to make it clear that we should not be claiming
compatibility with future APDB versions except at the patch level.