Write estimated visit footprint and time in preload.
Writing the visit properties in preload makes it possible to use them
in preprocessing pipelines. Both the footprint and the time have
limited accuracy, and will need to be improved later.

Note that the current ApPipe.yaml will call GetRegionTimeFromVisitTask
anyway until DM-43418 is done. This is acceptable; the outputs are
guaranteed to go in a different run, and the results (based on the
actual exposure metadata) are more accurate.