Warn non-standard FITS formatting but not raise an exception
The Rubin FITS files sometimes do not follow FITS standard fully.
While we read in the files and update a few fields, it can fail to
save due to astropy FITS verification errors from non-standard
cards that we do not modify. Just warn instead of raising an
exception so it does not stop processes.