Update guidelines on central repo versioning.
This formalizes our de-facto convention of having the unit test and
integration test repos bracket the range of Butler versions we support.

It also *inverts* previous guidance on /repo/embargo; we should be
checking in advance that we are compatible instead of letting
Middleware make an upgrade and then playing "catch-up".