Use regionTimeInfo's region to export template/refcat datasets
This changes the APIs of _export_skymap_and_templates and
_export_refcats to take a lsst.sphgeom.Region directly.

Note that the centroid of the region is not the same of the
detector center, but it should not matter.

Because htm7 can be too coarse compared to the patch size, using
htm7 indices to search for templates may lead to preloading more
patches than necessary and wasting time. This feature of using
htm7 to search for overlapping templates is also about to be
deprecated and replaced by the arbitrary spatial region query in
Butler. The usage will be replaced when switching to the new
butler query system.