Query one datasetType at a time for calibs/refcats/templates
This is a preparation step before switching to the new butler query
system.

In the new query system, query_datasets takes one datasetType at a
time, while butler.registry.queryDatasets can take a list of dataset
types in the old system. So, we change to query (via _filter_datasets)
one type of calibs/refcats/templates at a time.

Currently we can preload more types of calibs/refcats/templates
than the actual pipelines really need. It's possible that some types
are not preloaded but it's okay. For now we allow _MissingDatasetError.