Change _filter_datasets to take an iterable of dataset types
Querying butler one dataset type at a time is not necessary with
the butler.registry.queryDatasets. But this is a preparation step
before switching to the new query system which can only queries
one dataset type at a time.

Currently we can preload more types of calibs/refcats/templates
than the actual pipelines really need. It's possible that some types
are not preloaded but it's okay.