Support empty filter in nextVisit message.
The SAL spec allows filters to be the empty string to indicate that the
script did not specify a filter. If such a message is received, do not
constrain dataset queries by filter.

For this option to be useful, the dataset cache must be large enough to
store one calib per filter. The unit tests' default cache has been
expanded accordingly.