Split IgnorableVisit and InvalidVisitError.
We may want to reject nextVisit messages either because they specify
something we can't process (e.g., dome images for a differencing
pipeline) or because we specifically configured the service to ignore
them. The former is human error, the latter an error only in the formal
sense. Representing these cases with different exceptions lets them be
handled differently.