Warn instead of erroring on non-sky visits.
We routinely process non-sky visits such as daytime checkout to confirm
that the system is working before nightfall. However, doing so produces
error logs that get mixed in with processing or infrastructure errors.
Since a visit without a filter now emits a warning, a visit without a
position should do the same.