Do not skip group/exposure numbers in upload_from_repo.py
For every launch of the upload script, we used to skip
some numbers and leave a gap in the exposure IDs, for
easier output examination.

However the exposure ID space is more limited for LSSTCam-imSim.

For larger tests done via upload_from_repo.py, just use the next
possible number to conserve exposure IDs.

Not changhing upload.py as that test is much smaller typically.