Move Keda to FannedOutVisit conversion to message parsing.
This change lets FannedOutVisit be used as a universal message
representation in all other code. It also has the benefit that if the
message is invalid, the stream is not closed.

I considered moving the FAnnedOutVisit conversion into
RedisStreamSession, but this way keeps the class "clean" with only
Redis-specific elements.