Add support for template_coadd.
This is a standardized dataset type name that will be used going
forward. However, for now Prompt Processing must support both on a
pipeline-by-pipeline basis.

template_coadd is not yet covered by unit tests, because we don't have
any examples that can go into the local repo. We may need to switch
instruments to support this case.