Rename dataset query methods.
The old names of _export_[dataset category] are an artifact of when we
still used export/import, and are potentially confusing with pipeline
output export. Calling them _find_[category] is clearer.

While _find_refcats and _find_templates could be made more generic (the
only necessary difference is that one queries by HTM index and the
other by patch), I believe using specific collections instead of the
umbrella is an optimization worth keeping.