Remove gratuitous registry refreshes.
Refreshes are actually quite expensive (the cost is paid for by future
operations), and most of these are unnecessary now that init datasets
and output collections are initialized by a separate job.

The two remaining refreshes are both for the local repo: one after a
pipeline run (because the pipeline was run by a different Butler) and
one just before cleanup (better safe than sorry).