Remove flush_local_repo.
This function was used to recover from cases where a worker was killed
and restarted in the same pod. Since this scenario is not possible in
Keda (killing the worker kills the pod) and flush_local_repo had to
duplicate code with MiddlewareInterface.export_outputs, it's best
removed to avoid bit-rot.