Re-organize middleware export
To prepare for transferring datasets to the central Butler using Kafka, add an interface for an object that does the Butler writes back to central, to allow for alternate implementations.  This re-orders the middleware export process so that all of the writes to central happen at a single point in the code.