Separate out the raw copying to temp dir in upload_from_repo.py
Raws are copied to a temp dir before uploading to the central
repo because their metadata need to be modified.
The raw copying was done as part of the parallel upload after
the exposure.

This commit moves it out to a separate function before the
exposure, and provides options for it to go serially or
parallelly. It introduces a new waiting, as metadata modification
and uploading to the central repo only makes sense after the
corresponding raw is copied to the temp_dir.

This is a prep step before supporting using zipped raws where
all detector images of an exposure are stored in one zip file.
Before the zip format was introduced, each detector always had
its own fits file.