Start replayed exposures with a higher seq_num
This is to avoid conflicting the exposure IDs with the summit
generated exposures. They never get stored in the same butler
repo but having duplicate exposure IDs is a problem for
Sattle currently.

seq_num has a limit of 32768 for packing detector_exposure_id.
Currently some pipelines use it.  So seq_num cannot be too
high either.