Factor Knative/Gunicorn/Flask code into a separate module.
This removes dependencies on HTTP servers (of any kind) from the
activator. The largest coupling between the new driver_gunicorn and the
activator is in the set of exceptions that process_visit may raise.

This commit also removes the ability to run a Flask test server on a
local machine, a feature we haven't used since the earliest prototyping
and that requires a standalone installation of Flask.