Retry reading JSON sidecar with two approches
Previously, we only retry when the sidecar does not exist,
for occasional slower file transfer from summit.

But it's possible that the sidecar exists but is being
re-written and hence cannot be read temporarily.

We have seen intermittent ConcurrentModification errors
from the embargo Ceph S3. It could have been due to
summit re-sending successful transfer or other transient
storage issues. Hence, this retries in a longer timescale
when S3 ClientError is seen.

This also combines existence checking and file reading to
one operation so to avoid race.