Delegate "any pipelines?" check to MiddlewareInterface.
The check is done as soon after nextVisit receipt as possible to avoid
wasting resources on visits we don't care about. However, giving the
responsibility to MiddlewareInterface makes it possible to use cameras
or other Butler datasets in the decision.

As a side effect, this commit makes MiddlewareInterface's pipeline
config queries public.